Update game protocol to 13.16

+++ src/player.h
@@ -690,7 +690,7 @@ public:
 			client->sendCreatureShield(creature);
 		}
 	}
-	void sendSpellCooldown(uint8_t spellId, uint32_t time)
+	void sendSpellCooldown(uint16_t spellId, uint32_t time)
 	{
 		if (client) {
 			client->sendSpellCooldown(spellId, time);
